+++ RNDeviceInfo/RNDeviceInfo.m
@@ -11,29 +11,38 @@
 #import <LocalAuthentication/LocalAuthentication.h>
 
 @interface RNDeviceInfo()
-
+@property (nonatomic) bool isEmulator;
 @end
 
 @implementation RNDeviceInfo
-{
 
-}
+@synthesize isEmulator;
 
 RCT_EXPORT_MODULE()
-
-- (dispatch_queue_t)methodQueue
+    
++ (BOOL)requiresMainQueueSetup
 {
-    return dispatch_get_main_queue();
+   return YES;
 }
+
 
 - (NSString*) deviceId
 {
     struct utsname systemInfo;
 
     uname(&systemInfo);
-
-    return [NSString stringWithCString:systemInfo.machine
-                                    encoding:NSUTF8StringEncoding];
+    
+    NSString* deviceId = [NSString stringWithCString:systemInfo.machine
+                                            encoding:NSUTF8StringEncoding];
+    
+    if ([deviceId isEqualToString:@"i386"] || [deviceId isEqualToString:@"x86_64"] ) {
+        deviceId = [NSString stringWithFormat:@"%s", getenv("SIMULATOR_MODEL_IDENTIFIER")];
+        self.isEmulator = YES;
+    } else {
+        self.isEmulator = NO;
+    }
+    
+    return deviceId;
 }
 
 - (NSString*) deviceName
@@ -41,10 +50,8 @@ RCT_EXPORT_MODULE()
     static NSDictionary* deviceNamesByCode = nil;
 
     if (!deviceNamesByCode) {
-
-        deviceNamesByCode = @{@"i386"      :@"Simulator",
-                              @"x86_64"    :@"Simulator",
-                              @"iPod1,1"   :@"iPod Touch",      // (Original)
+        
+        deviceNamesByCode = @{@"iPod1,1"   :@"iPod Touch",      // (Original)
                               @"iPod2,1"   :@"iPod Touch",      // (Second Generation)
                               @"iPod3,1"   :@"iPod Touch",      // (Third Generation)
                               @"iPod4,1"   :@"iPod Touch",      // (Fourth Generation)
@@ -86,6 +93,12 @@ RCT_EXPORT_MODULE()
                               @"iPhone9,3" :@"iPhone 7",        // (model A1778 | Global)
                               @"iPhone9,2" :@"iPhone 7 Plus",   // (model A1661 | CDMA)
                               @"iPhone9,4" :@"iPhone 7 Plus",   // (model A1784 | Global)
+                              @"iPhone10,3":@"iPhone X",        // (model A1865, A1902)
+                              @"iPhone10,6":@"iPhone X",        // (model A1901)
+                              @"iPhone10,1":@"iPhone 8",        // (model A1863, A1906, A1907)
+                              @"iPhone10,4":@"iPhone 8",        // (model A1905)
+                              @"iPhone10,2":@"iPhone 8 Plus",   // (model A1864, A1898, A1899)
+                              @"iPhone10,5":@"iPhone 8 Plus",   // (model A1897)
                               @"iPad4,1"   :@"iPad Air",        // 5th Generation iPad (iPad Air) - Wifi
                               @"iPad4,2"   :@"iPad Air",        // 5th Generation iPad (iPad Air) - Cellular
                               @"iPad4,3"   :@"iPad Air",        // 5th Generation iPad (iPad Air)
@@ -103,10 +116,15 @@ RCT_EXPORT_MODULE()
                               @"iPad6,4"   :@"iPad Pro 9.7-inch",// iPad Pro 9.7-inch
                               @"iPad6,7"   :@"iPad Pro 12.9-inch",// iPad Pro 12.9-inch
                               @"iPad6,8"   :@"iPad Pro 12.9-inch",// iPad Pro 12.9-inch
+                              @"iPad7,1"   :@"iPad Pro 12.9-inch",// 2nd Generation iPad Pro 12.5-inch - Wifi
+                              @"iPad7,2"   :@"iPad Pro 12.9-inch",// 2nd Generation iPad Pro 12.5-inch - Cellular
+                              @"iPad7,3"   :@"iPad Pro 10.5-inch",// iPad Pro 12.5-inch - Wifi
+                              @"iPad7,4"   :@"iPad Pro 10.5-inch",// iPad Pro 12.5-inch - Cellular
                               @"AppleTV2,1":@"Apple TV",        // Apple TV (2nd Generation)
                               @"AppleTV3,1":@"Apple TV",        // Apple TV (3rd Generation)
                               @"AppleTV3,2":@"Apple TV",        // Apple TV (3rd Generation - Rev A)
                               @"AppleTV5,3":@"Apple TV",        // Apple TV (4th Generation)
+                              @"AppleTV6,2":@"Apple TV 4K",     // Apple TV 4K
                               };
     }
 
@@ -124,6 +142,9 @@ RCT_EXPORT_MODULE()
         else if([self.deviceId rangeOfString:@"iPhone"].location != NSNotFound){
             deviceName = @"iPhone";
         }
+        else if([self.deviceId rangeOfString:@"AppleTV"].location != NSNotFound){
+            deviceName = @"Apple TV";
+        }
     }
 
     return deviceName;
@@ -131,8 +152,12 @@ RCT_EXPORT_MODULE()
 
 - (NSString*) userAgent
 {
+#if TARGET_OS_TV
+    return @"not available";
+#else
     UIWebView* webView = [[UIWebView alloc] initWithFrame:CGRectZero];
     return [webView stringByEvaluatingJavaScriptFromString:@"navigator.userAgent"];
+#endif
 }
 
 - (NSString*) deviceLocale
@@ -153,11 +178,6 @@ RCT_EXPORT_MODULE()
   return currentTimeZone.name;
 }
 
-- (bool) isEmulator
-{
-  return [self.deviceName isEqual: @"Simulator"];
-}
-
 - (bool) isTablet
 {
   return [[UIDevice currentDevice] userInterfaceIdiom] == UIUserInterfaceIdiomPad;
@@ -172,6 +192,7 @@ RCT_EXPORT_MODULE()
     return @{
              @"systemName": currentDevice.systemName,
              @"systemVersion": currentDevice.systemVersion,
+             @"apiLevel": @"not available",
              @"model": self.deviceName,
              @"brand": @"Apple",
              @"deviceId": self.deviceId,